+++ repository/get_book_request.go
@@ -1,0 +1,5 @@
+package repository
+
+type GetBookRequest struct {
+	ID int64
+}
